refactor: update frontend services nomenclature
