+++ packages/frontend/src/services/adapters/authState.adapter.ts
@@ -1,0 +1,11 @@
+import { EndpointAuth } from '@Interfaces/ports/endpoint/endpoint-auth-state.interface';
+
+export const createAdaptedAuthState = (resAuthState: EndpointAuth) => {
+  const { user, jwt } = resAuthState;
+  const formattedAuthState = {
+    user,
+    jwt,
+  };
+
+  return formattedAuthState;
+};
