+++ packages/frontend/src/services/notification.service.ts
@@ -1,5 +1,5 @@
 import { toast } from 'react-toastify';
-import type { NotificationService } from '@Interfaces/ports/notification.interface';
+import type { NotificationService } from '@Interfaces/ports/service/notification-service.interface';
 
 export function useNotifier(): NotificationService {
   const notifyError = ({ message }: { message: string }) => {
